+++ php/getUsers.php
@@ -1,31 +1,59 @@
 <?php
 include("php/config.php");
-$mail = $_SESSION['mail'];
-$req1 = $conn->prepare("select * from utilisateur where mail != '$mail'");
-$req1 -> execute();
 
-$select = '';
+if(isset($_SESSION['grade'])){
 
-if($req1 -> rowCount() > 0){
-    while($data1 = $req1 -> fetch(PDO::FETCH_ASSOC)){
-        $PseudoTmp = $data1['pseudo'];
-        $MailTmp = $data1['mail'];
-        $select .= '<li class="list-group-item">' .$PseudoTmp .'</br>'.$MailTmp .'
-                <form method ="post" class="form-check-inline" style="float:right;">
-                    <button style="float:right;" name="suppr" value='.$MailTmp.'> supprimer </button>
-                </form>
-            </li>';
+    $sess = $_SESSION['grade'];
+    $mail = $_SESSION['mail'];
+    $Theme ='';
+
+    if($sess==3){
+        $req1 = $conn->prepare("select * from utilisateur where mail != '$mail'");
+        $Theme ='Liste des thèmes';
+    }else{
+        $req1 = $conn->prepare("select * from utilisateur where mail != '$mail' and role = 1 ");
     }
-}
 
-if (isset($_POST['suppr'])) {
-    $mailASup = $_POST['suppr'];
-    $req1 = $conn->prepare("delete from utilisateur where mail = '$mailASup'");
+
     $req1 -> execute();
-    header("Location: displayAdmin.php");
+
+    $select = '<div class="container.fluid" id = themeUser>
+    <div class="row">
+            <div class="col-6" style="color:white;">
+                <h2>Liste des Utilisateurs</h2>
+            </div>
+            <div class="col-6" style="color:white;">
+                <h2>'.$Theme .'</h2>
+            </div>
+        </div>
+    </div>
+    <div class="container.fluid" id = themeUser>
+        <div class="row">
+            <div class="col-6 pb-4">
+                <ul class="list-group">';
+
+
+
+    if($req1 -> rowCount() > 0){
+        while($data1 = $req1 -> fetch(PDO::FETCH_ASSOC)){
+            $PseudoTmp = $data1['pseudo'];
+            $MailTmp = $data1['mail'];
+            $select .= '<li class="list-group-item">' .$PseudoTmp .'</br>'.$MailTmp .'
+                    <form method ="post" class="form-check-inline" style="float:right;">
+                        <button style="float:right;" name="suppr" value='.$MailTmp.'> supprimer </button>
+                    </form>
+                </li>';
+        }
+    }
+
+    if (isset($_POST['suppr'])) {
+        $mailASup = $_POST['suppr'];
+        $req1 = $conn->prepare("delete from utilisateur where mail = '$mailASup'");
+        $req1 -> execute();
+        header("Location: displayAdmin.php");
+    }
+
+    echo $select;
 }
-
-
-echo $select;
 $conn = null;
 ?>
